Change production URL Change version number to single digit, and increment Rename the SDK assigned sessionId to internalSessionId Rename event provided gameSession to session (externally visible and now in conformance w/ other SDKs) Fixes issues w/ appId sometimes long, now all signed long long Include transactionId in URL for Transaction event submission Update copyright header on externally distributed file Remove userInfo event w/ no parameters Changed types:   SessionStart, SessionEnd, GameStart, GameEnd: sessionId NSString => signed long long   GameStart, GameEnd: instance NSString => signed long long   InvitationSent, InvitationResponse: invitationId NSString => signed long long   Transaction: transactionId NSString => signed long long Adjusted build parameters Probably some other changes

+++ iosapi/iosapi/PNTransactionEvent.h
@@ -1,7 +1,7 @@
 #import "PNEvent.h"
 
 @interface PNTransactionEvent : PNEvent {
-    long _transactionId;
+    signed long long _transactionId;
     NSString * _itemId;
     double _quantity;
     PNTransactionType _type;
@@ -11,7 +11,7 @@
     NSArray * _currencyCategories;
 }
 
-@property(nonatomic, assign) long transactionId;
+@property(nonatomic, assign) signed long long transactionId;
 @property(nonatomic, retain) NSString * itemId;
 @property(nonatomic, assign) double quantity;
 @property(nonatomic, assign) PNTransactionType type;
@@ -26,9 +26,9 @@
  *  currencyCategories: Array of PNCurrencyCategory NSNumbers
  */
 - (id) init:  (PNEventType) eventType 
-              applicationId: (long) applicationId 
+              applicationId: (signed long long) applicationId
                      userId: (NSString *) userId 
-              transactionId: (long) transactionId 
+              transactionId: (signed long long) transactionId
                      itemId: (NSString *) itemId 
                    quantity: (double) quantity 
                        type: (PNTransactionType) type
